Split queries and mutations
