+++ src/api/mutations/commonFragments.js
@@ -1,0 +1,8 @@
+import { gql } from "@apollo/client";
+
+export const Response = gql`
+  fragment Response on Answer {
+    status
+    message
+  }
+`;
