+++ src/api/mutations/user.js
@@ -1,0 +1,7 @@
+import { gql } from "@apollo/client";
+
+export const LOGIN = gql`
+  mutation login($email: String!, $password: String!) {
+    login(email: $email, password: $password)
+  }
+`;
