+++ src/api/queries/favoriteGenre.js
@@ -1,0 +1,7 @@
+import { gql } from "@apollo/client";
+
+export const GET_FAVORITE_GENRES = gql`
+  query getFavoriteGenres {
+    favoriteGenres
+  }
+`;
