+++ src/api/queries/favoriteMovie.js
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client";
+
+export const GET_FAVORITE_MOVIES = gql`
+  query getFavoriteMovies {
+    favoriteMovies {
+      movieId
+      isViewed
+    }
+  }
+`;
