+++ src/api/queries/genre.js
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client";
+
+export const GET_GENRES = gql`
+  query getGenres($language: String!) {
+    genres(language: $language) {
+      id
+      name
+    }
+  }
+`;
