+++ src/api/queries/movie.js
@@ -1,27 +1,5 @@
 import { gql } from "@apollo/client";
 
-export const GET_USER = gql`
-  query getUser {
-    user {
-      userId
-      email
-      password
-    }
-  }
-`;
-export const GET_GENRES = gql`
-  query getGenres($language: String!) {
-    genres(language: $language) {
-      id
-      name
-    }
-  }
-`;
-export const GET_FAVORITE_GENRES = gql`
-  query getFavoriteGenres {
-    favoriteGenres
-  }
-`;
 const MOVIE_INFO = gql`
   fragment MovieInfo on Movie {
     id
@@ -45,14 +23,6 @@ export const GET_MOVIE_BY_ID = gql`
     }
   }
   ${MOVIE_INFO}
-`;
-export const GET_FAVORITE_MOVIES = gql`
-  query getFavoriteMovies {
-    favoriteMovies {
-      movieId
-      isViewed
-    }
-  }
 `;
 export const GET_FILTERED_MOVIES = gql`
   query getFilteredMovies(
