+++ src/api/queries/user.js
@@ -1,0 +1,11 @@
+import { gql } from "@apollo/client";
+
+export const GET_USER = gql`
+  query getUser {
+    user {
+      userId
+      email
+      password
+    }
+  }
+`;
