+++ src/components/hooks/useFavoriteGenres.js
@@ -1,5 +1,5 @@
 import { useQuery } from "@apollo/client";
-import { GET_FAVORITE_GENRES } from "apolloClient/queries";
+import { GET_FAVORITE_GENRES } from "api/queries/favoriteGenre";
 
 export default function useFavoriteGenres() {
   const { loading, data, refetch } = useQuery(GET_FAVORITE_GENRES);
