+++ src/components/hooks/useFavoriteMovies.js
@@ -1,5 +1,5 @@
 import { useQuery } from "@apollo/client";
-import { GET_FAVORITE_MOVIES } from "apolloClient/queries";
+import { GET_FAVORITE_MOVIES } from "api/queries/favoriteMovie";
 
 export default function useFavoriteMovies() {
   const { loading, data, refetch } = useQuery(GET_FAVORITE_MOVIES);
