+++ src/components/hooks/useFavoriteMoviesMutation.js
@@ -1,5 +1,5 @@
 import { useMutation } from "@apollo/client";
-import { SET_FAVORITE_MOVIES } from "apolloClient/mutations";
+import { SET_FAVORITE_MOVIES } from "api/mutations/favoriteMovie";
 
 export default function useFavoriteMoviesMutation(onCompleted, movieId) {
   const [setFavoriteMovies] = useMutation(SET_FAVORITE_MOVIES, {
