+++ src/components/hooks/useGenres.js
@@ -1,5 +1,5 @@
 import { useQuery } from "@apollo/client";
-import { GET_GENRES } from "apolloClient/queries";
+import { GET_GENRES } from "api/queries/genre";
 
 export default function useGenres(language = "en-US") {
   const { loading, data } = useQuery(GET_GENRES, {
